+++ packages/at_secondary_server/lib/src/verb/handler/info_verb_handler.dart
@@ -2,9 +2,7 @@ import 'dart:collection';
 import 'dart:convert';
 
 import 'package:at_commons/at_commons.dart';
-import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
 import 'package:at_secondary/src/connection/inbound/inbound_connection_metadata.dart';
-import 'package:at_secondary/src/constants/enroll_constants.dart';
 import 'package:at_secondary/src/server/at_secondary_config.dart';
 import 'package:at_secondary/src/server/at_secondary_impl.dart';
 import 'package:at_server_spec/at_server_spec.dart';
@@ -33,11 +31,8 @@ class InfoVerbHandler extends AbstractVerbHandler {
       HashMap<String, String?> verbParams,
       InboundConnection atConnection) async {
     Map infoMap = {};
-    String? apkamMetadataKey;
-    String? result;
     InboundConnectionMetadata atConnectionMetadata = atConnection.metaData
         as InboundConnectionMetadata; // structure of what is returned is documented in the [Info] verb in at_server_spec
-    var atSign = AtSecondaryServerImpl.getInstance().currentAtSign;
 
     infoMap['version'] = AtSecondaryConfig.secondaryServerVersion;
     Duration uptime = Duration(
@@ -46,14 +41,11 @@ class InfoVerbHandler extends AbstractVerbHandler {
     if (verbParams[paramFullCommandAsReceived] == 'info') {
       String uptimeAsWords = durationToWords(uptime);
       infoMap['uptimeAsWords'] = uptimeAsWords;
-      final enrollApprovalId = atConnectionMetadata.enrollmentId;
-      if (atConnectionMetadata.isAuthenticated && enrollApprovalId != null) {
-        apkamMetadataKey =
-            '$enrollApprovalId.$newEnrollmentKeyPattern.$enrollManageNamespace$atSign';
-        result = await _getApkamMetadataKey(apkamMetadataKey);
-        if (result != null) {
-          infoMap['apkam_metadata'] = result;
-        }
+      if (atConnectionMetadata.isAuthenticated &&
+          atConnectionMetadata.enrollmentId != null) {
+        infoMap['apkam_metadata'] = await AtSecondaryServerImpl.getInstance()
+            .enrollmentManager
+            .get(atConnectionMetadata.enrollmentId!);
       }
     } else {
       infoMap['uptimeAsMillis'] = uptime.inMilliseconds;
@@ -72,15 +64,5 @@ class InfoVerbHandler extends AbstractVerbHandler {
         ((uDays > 0 || uHours > 0 || uMins > 0) ? "$uMins minutes " : "") +
         "$uSeconds seconds";
     return uptimeAsWords;
-  }
-
-  Future<String?> _getApkamMetadataKey(String? apkamMetadataKey) async {
-    AtData? result;
-    try {
-      result = await keyStore.get(apkamMetadataKey);
-    } on KeyNotFoundException {
-      logger.warning('apkam key $apkamMetadataKey not found');
-    }
-    return result?.data;
   }
 }