+++ packages/at_secondary_server/CHANGELOG.md
@@ -2,8 +2,11 @@
 - feat: Introduce option to unrevoke revoked enrollments
 - feat: Introduce option to delete enrollments that are denied/revoked
 - build[deps]: update dependency versions of at_commons -> 4.1.2, at_utils -> 3.0.18, at_lookup -> 3.0.48
+- fix: LatestCommitEntryOfEachKey metric fixed to return commit log entries till last commitID instead of default limit 25.
 ## 3.0.50
 - fix: Enhance namespace authorisation check to verify when namespace has a period in it
+- feat: Enable expiration of APKAM keys based on the specified duration.
+
 ## 3.0.49
 - feat: Enforce superset access check for approving apps
 - fix: respect isEncrypted:false if supplied in the notify: command, and 
@@ -13,6 +16,7 @@
   notifications from the server to the client.
 - build[deps]: Upgraded the following package:
   - at_persistence_secondary_server to v3.0.63
+
 ## 3.0.48
 - feat Add expiresAt and availableAt params to notify:list response